+++ src/components/lab/Autocomplete/story/index.jsx
@@ -5,7 +5,7 @@ import { Autocomplete } from '../Autocomplete'
 const page = PicassoBook.createPage(
   'Autocomplete',
   `Input with the autocomplete`,
-  'Forms'
+  'Lab'
 )
 
 page.createTabChapter('Props').addComponentDocs({
@@ -41,8 +41,15 @@ for this component. This you can achieve by adding corresponding attributes:
   .addExample('lab/Autocomplete/story/AllowAnyDisabled.example.jsx', {
     title: 'Any value disabled',
     description: `By default Autocomplete allows any entered input value to stay after focus is removed from input,
-        you can set \`allowAny={false}\` prop to allow only selected values to stay after focus is removed from input`
-  })
+      you can set \`allowAny={false}\` prop to allow only selected values to stay after focus is removed from input`
+  }) // picasso-skip-visuals
+  .addExample('lab/Autocomplete/story/OtherOption.example.jsx', {
+    title: 'Other option',
+    description: `By default Autocomplete allows any entered input value to stay after focus is removed from input,
+      you can set \`showOtherOption={true}\` prop to allow also this new item to appear in the suggestions list. Also,
+      you can decorate this option with the prefix text by using \`otherOptionText\` and handle selection with
+      \`onOtherOptionSelect\` event handler.`
+  }) // picasso-skip-visuals
   .addExample(
     'lab/Autocomplete/story/InitialSelectedItem.example.jsx',
     'Initially selected item'